RestAPI Exception Logic Enhancement

+++ src/main/java/com/study/dawn/controller/BoardApiController.java
@@ -5,6 +5,9 @@ import com.study.dawn.exception.NotFoundException;
 import com.study.dawn.service.BoardService;
 import com.study.dawn.vo.BoardVO;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import java.util.HashMap;
 
@@ -16,12 +19,16 @@ public class BoardApiController {
     private BoardService boardService;
 
     @PostMapping
-    public void insertPost(@RequestBody BoardVO boardVO) throws Exception{
+    public ResponseEntity insertPost(@RequestBody BoardVO boardVO) throws Exception{
 
         try {
             String title = boardVO.getTitle();
             String content = boardVO.getContent();
             String writer = boardVO.getWriter();
+
+            if(StringUtils.isEmpty(title) || StringUtils.isEmpty(content) || StringUtils.isEmpty(writer)){
+                throw new NotFoundException(ErrorCode.INTERNAL_SERVER_ERROR, "Please enter a required value.");
+            }
 
             HashMap<String, Object> map = new HashMap<String, Object>();
             map.put("title", title);
@@ -29,6 +36,9 @@ public class BoardApiController {
             map.put("writer", writer);
 
             boardService.regist(map);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+
         }catch(NotFoundException ne){
             throw new NotFoundException(ErrorCode.INTERNAL_SERVER_ERROR, ne.getMessage(), ne);
         }catch(Exception e){
@@ -37,33 +47,46 @@ public class BoardApiController {
     }
 
     @GetMapping("/{bno}")
-    public HashMap<String, Object> readPost(@PathVariable("bno") int bno) throws Exception{
+    public ResponseEntity readPost(@PathVariable("bno") int bno) throws Exception{
 
-        HashMap<String, Object> map = new HashMap<String, Object>();
         try {
-            map = boardService.read(bno);
+            if(StringUtils.isEmpty(bno)){
+                throw new NotFoundException(ErrorCode.INTERNAL_SERVER_ERROR, "Please enter a required value.");
+            }
+
+            HashMap<String, Object> map  = boardService.read(bno);
+
+            if(StringUtils.isEmpty(map)){
+                throw new NotFoundException(ErrorCode.INTERNAL_SERVER_ERROR, "No data");
+            }
+            return new ResponseEntity<>(map, HttpStatus.OK);
 
         }catch(NotFoundException ne){
             throw new NotFoundException(ErrorCode.INTERNAL_SERVER_ERROR, ne.getMessage(), ne);
         }catch(Exception e){
             throw new Exception(e.getMessage(), e);
         }
-        return map;
     }
 
     @PutMapping("/{bno}")
-    public void updatePost(@PathVariable int bno, @RequestBody BoardVO boardVO) throws Exception{
+    public ResponseEntity updatePost(@PathVariable int bno, @RequestBody BoardVO boardVO) throws Exception{
 
         try {
             String title = boardVO.getTitle();
             String content = boardVO.getContent();
 
+            if(StringUtils.isEmpty(title) || StringUtils.isEmpty(content)){
+                throw new NotFoundException(ErrorCode.INTERNAL_SERVER_ERROR, "Please enter a required value.");
+            }
+
             HashMap<String, Object> map = new HashMap<String, Object>();
             map.put("bno", bno);
             map.put("title", title);
             map.put("content", content);
-
             boardService.modify(map);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+
         }catch(NotFoundException ne){
             throw new NotFoundException(ErrorCode.INTERNAL_SERVER_ERROR, ne.getMessage(), ne);
         }catch(Exception e){
@@ -72,10 +95,14 @@ public class BoardApiController {
     }
 
     @DeleteMapping("/{bno}")
-    public void deletePost(@PathVariable("bno") int bno) throws Exception{
+    public ResponseEntity deletePost(@PathVariable("bno") int bno) throws Exception{
 
         try {
+            if(StringUtils.isEmpty(bno)){
+                throw new NotFoundException(ErrorCode.INTERNAL_SERVER_ERROR, "Please enter a required value.");
+            }
             boardService.remove(bno);
+            return new ResponseEntity<>(HttpStatus.OK);
         }catch(NotFoundException ne){
             throw new NotFoundException(ErrorCode.INTERNAL_SERVER_ERROR, ne.getMessage(), ne);
         }catch(Exception e){
